pyproject.toml
[project]
name = "lexbridge"
version = "0.1.0"
description = "LexBridge: Cross-Lingual M&A Due Diligence Intelligence - Fully Local Multi-Agent System powered by Microsoft Harrier embeddings"
requires-python = ">=3.12,<3.13"
license = "MIT"
readme = "README.md"
dependencies = [
    "numpy<2",
    "torch>=2.2",
    "transformers>=4.51",
    "sentence-transformers>=3.4",
    "tokenizers>=0.21",
    "qdrant-client>=1.12",
    "falkordblite>=0.9",
    "langgraph>=0.3",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.3",
    "streamlit>=1.40",
    "pyvis>=0.3",
    "plotly>=5.0",
    "langdetect>=1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "huggingface-hub>=0.25",
    "streamlit-agraph>=0.0.45",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
